+++ src/main/java/com/actiontech/dble/config/model/user/ShardingUserConfig.java
@@ -15,8 +15,8 @@ public class ShardingUserConfig extends ServerUserConfig {
     private Set<String> schemas;
     private UserPrivilegesConfig privilegesConfig;
 
-    public ShardingUserConfig(String name, String password, String strWhiteIPs, int maxCon) {
-        super(name, password, strWhiteIPs, maxCon);
+    public ShardingUserConfig(String name, String password, String strWhiteIPs, String strMaxCon) {
+        super(name, password, strWhiteIPs, strMaxCon);
     }
