+++ src/main/java/com/actiontech/dble/config/model/user/UserConfig.java
@@ -6,6 +6,7 @@
 package com.actiontech.dble.config.model.user;
 
 import com.actiontech.dble.util.SplitUtil;
+import com.actiontech.dble.util.StringUtil;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -17,11 +18,19 @@ public class UserConfig {
     protected final Set<String> whiteIPs;
     protected int maxCon;
 
-    public UserConfig(String name, String password, String strWhiteIPs, int maxCon) {
+    public UserConfig(String name, String password, String strWhiteIPs, String strMaxCon) {
         this.name = name;
         this.password = password;
         this.whiteIPs = genWhiteIPs(strWhiteIPs);
-        this.maxCon = maxCon;
+
+        int maxConn = -1;
+        if (!StringUtil.isEmpty(strMaxCon)) {
+            maxConn = Integer.parseInt(strMaxCon);
+            if (maxConn < 0) {
+                maxConn = -1;
+            }
+        }
+        this.maxCon = maxConn;
     }
 
     private Set<String> genWhiteIPs(String strWhiteIPs) {
@@ -50,9 +59,5 @@ public class UserConfig {
 
     public int getMaxCon() {
         return maxCon;
-    }
-
-    public void setMaxCon(int maxCon) {
-        this.maxCon = maxCon;
     }
 }
