+++ src/main/resources/user_template.xml
@@ -14,7 +14,7 @@
 	under the License. -->
 <!DOCTYPE dble:user SYSTEM "user.dtd">
 <dble:user xmlns:dble="http://dble.cloud/" version="4.0">
-    <managerUser name="man1" password="654321"/>
+    <managerUser name="man1" password="654321" maxCon="100"/>
     <managerUser name="user" usingDecrypt="true" whiteIPs="127.0.0.1,0:0:0:0:0:0:0:1" readOnly="true" password="AqEkFEuIFAX6g2TJQnp4cJ2r7Yc0Z4/KBsZqKhT8qSz18Aj91e8lxO49BKQElC6OFfW4c38pCYa8QGFTub7pnw==" />
 
     <shardingUser name="root" password="123456" schemas="testdb" readOnly="false" blacklist="blacklist1" maxCon="20"/>
